move response types to neutral location for api
