+++ src/rest_api/api/compound_processing/mod.rs
@@ -1,6 +1,3 @@
-mod compound_processing_management;
-pub use compound_processing_management::*;
-
 mod convert_mol_block_to_smiles;
 pub use convert_mol_block_to_smiles::*;
