+++ src/rest_api/api/mod.rs
@@ -5,4 +5,8 @@ mod search;
 pub use search::*;
 
 mod compound_processing;
+
+mod response_types;
+pub use response_types::*;
+
 pub use compound_processing::*;
