+++ src/rest_api/api/response_types.rs
@@ -1,83 +1,148 @@
+use crate::search::{QuerySearchHit, StructureSearchHit};
 use poem_openapi::{payload::Json, ApiResponse, Object};
 use tantivy::Opstamp;
+
+// Response types
+#[derive(ApiResponse, Debug)]
+pub enum StandardizeResponse {
+    #[oai(status = "200", content_type = "application/json")]
+    Ok(Json<Vec<StandardizedSmiles>>),
+}
+
+#[derive(ApiResponse, Debug)]
+pub enum ConvertedSmilesResponse {
+    #[oai(status = "200", content_type = "application/json")]
+    Ok(Json<Vec<ConvertedSmiles>>),
+}
+
+#[derive(ApiResponse, Debug)]
+pub enum ConvertedMolBlockResponse {
+    #[oai(status = "200", content_type = "application/json")]
+    Ok(Json<Vec<ConvertedMolBlock>>),
+}
 
 #[derive(ApiResponse, Debug)]
 pub enum GetIndexResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<IndexSchema>),
+    Ok(Json<crate::rest_api::api::IndexSchema>),
     #[oai(status = "400")]
     NotFound,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<GetIndexesResponseError>),
+    Err(Json<crate::rest_api::api::GetIndexesResponseError>),
 }
 
 #[derive(ApiResponse)]
 pub enum GetIndexesResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<Vec<IndexMeta>>),
+    Ok(Json<Vec<crate::rest_api::api::IndexMeta>>),
     #[oai(status = "400")]
     NotFound,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<GetIndexesResponseError>),
+    Err(Json<crate::rest_api::api::GetIndexesResponseError>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum ListIndexesResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<Vec<IndexMeta>>),
+    Ok(Json<Vec<crate::rest_api::api::IndexMeta>>),
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<ListIndexResponseErr>),
+    Err(Json<crate::rest_api::api::ListIndexResponseErr>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum ListSchemasResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<Vec<Schema>>),
+    Ok(Json<Vec<crate::rest_api::api::Schema>>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum PostIndexesBulkIndexResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<PostIndexBulkResponseOk>),
+    Ok(Json<crate::rest_api::api::PostIndexBulkResponseOk>),
     #[oai(status = "404")]
     IndexDoesNotExist,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<PostIndexBulkResponseError>),
+    Err(Json<crate::rest_api::api::PostIndexBulkResponseError>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum DeleteIndexesBulkDeleteResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<DeleteIndexBulkResponseOk>),
+    Ok(Json<crate::rest_api::api::DeleteIndexBulkResponseOk>),
     #[oai(status = "404")]
     IndexDoesNotExist,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<DeleteIndexBulkResponseError>),
+    Err(Json<crate::rest_api::api::DeleteIndexBulkResponseError>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum PostIndexResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<IndexMeta>),
+    Ok(Json<crate::rest_api::api::IndexMeta>),
     #[oai(status = "404")]
     SchemaDoesNotExist,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<CreateIndexError>),
+    Err(Json<crate::rest_api::api::CreateIndexError>),
 }
 
 #[derive(ApiResponse, Debug)]
 pub enum DeleteIndexResponse {
     #[oai(status = "200", content_type = "application/json")]
-    Ok(Json<IndexMeta>),
+    Ok(Json<crate::rest_api::api::IndexMeta>),
     #[oai(status = "404")]
     IndexDoesNotExist,
     #[oai(status = "500", content_type = "application/json")]
-    Err(Json<DeleteIndexError>),
+    Err(Json<crate::rest_api::api::DeleteIndexError>),
+}
+
+#[derive(ApiResponse, Debug)]
+pub enum GetQuerySearchResponse {
+    #[oai(status = "200", content_type = "application/json")]
+    Ok(Json<Vec<QuerySearchHit>>),
+    #[oai(status = "404")]
+    IndexDoesNotExist,
+    #[oai(status = "500", content_type = "application/json")]
+    Err(Json<crate::rest_api::api::QueryResponseError>),
+}
+
+#[derive(ApiResponse, Debug)]
+pub enum GetStructureSearchResponse {
+    #[oai(status = "200", content_type = "application/json")]
+    Ok(Json<Vec<StructureSearchHit>>),
+    #[oai(status = "404")]
+    IndexDoesNotExist,
+    #[oai(status = "500", content_type = "application/json")]
+    Err(Json<crate::rest_api::api::StructureResponseError>),
+}
+
+// Relevant structs
+#[derive(Object, Debug)]
+pub struct StandardizedSmiles {
+    #[oai(skip_serializing_if_is_none)]
+    pub smiles: Option<String>,
+    #[oai(skip_serializing_if_is_none)]
+    pub error: Option<String>,
+}
+
+#[derive(Object, Debug)]
+pub struct ConvertedSmiles {
+    #[oai(skip_serializing_if_is_none)]
+    pub smiles: Option<String>,
+    #[oai(skip_serializing_if_is_none)]
+    pub error: Option<String>,
+}
+
+#[derive(Object, Debug)]
+pub struct ConvertedMolBlock {
+    #[oai(skip_serializing_if_is_none)]
+    pub mol_block: Option<String>,
+    #[oai(skip_serializing_if_is_none)]
+    pub error: Option<String>,
 }
 
 #[derive(Object, Debug)]
 pub struct BulkRequest {
-    pub docs: Vec<BulkRequestDoc>,
+    pub docs: Vec<crate::rest_api::api::BulkRequestDoc>,
 }
 
 #[derive(Object, Debug)]
@@ -126,7 +191,7 @@ pub struct PostIndexBulkResponseError {
 
 #[derive(Object, Debug)]
 pub struct PostIndexBulkResponseOk {
-    pub statuses: Vec<PostIndexBulkResponseOkStatus>,
+    pub statuses: Vec<crate::rest_api::api::PostIndexBulkResponseOkStatus>,
 }
 
 #[derive(Object, Debug)]
@@ -142,7 +207,7 @@ pub struct DeleteIndexBulkResponseError {
 
 #[derive(Object, Debug)]
 pub struct DeleteIndexBulkResponseOk {
-    pub statuses: Vec<DeleteIndexBulkResponseOkStatus>,
+    pub statuses: Vec<crate::rest_api::api::DeleteIndexBulkResponseOkStatus>,
 }
 
 #[derive(Object, Debug)]
@@ -155,4 +220,14 @@ pub struct DeleteIndexBulkResponseOkStatus {
 pub struct Schema {
     pub name: String,
     pub schema: serde_json::Value,
+}
+
+#[derive(Object, Debug)]
+pub struct QueryResponseError {
+    pub error: String,
+}
+
+#[derive(Object, Debug)]
+pub struct StructureResponseError {
+    pub error: String,
 }
