+++ src/rest_api/api/search/mod.rs
@@ -1,9 +1,6 @@
 mod basic_search;
 pub use basic_search::*;
 
-mod search_management;
-pub use search_management::*;
-
 mod identity_search;
 pub use identity_search::*;
